backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.1.2
python-multipart
PyJWT
slowapi
python-dotenv